extract pipe to Map function and test it
